Daniel

+++ app/Http/Controllers/AgendaController.php
@@ -24,10 +24,15 @@ class AgendaController extends Controller
     private function userInauthenticated()
     {
         $this->user = $this->service->getUserAuthenticated();
-        if (!$this->user || $this->user->rol->nombre != 'Empresa') {
-            return view('auth/login', ['message' => 'No se ha logueado o no tiene los permisos']);
-        }
+        if (!$this->user) {
+            return view('auth/login', ['message' => 'No se ha logueado']);
+        } else if ($this->user->rol->name != 'Administrador') {
+            return view('home', ['message' => 'No tiene los permisos para ejecutar esta acción']);
+        } 
+        
+        return null;
     }
+
     private function translateDate($date) {
         $timestamp = strtotime($date);
 
@@ -73,7 +78,8 @@ class AgendaController extends Controller
      */
     public function index()
     {
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
         
         $agendas= Agenda::with('place', 'stand')->get();
         
@@ -103,7 +109,8 @@ class AgendaController extends Controller
     public function create()
     {
 
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
 
         $stands= Stand::all();
         $places= Places::all();
@@ -119,7 +126,8 @@ class AgendaController extends Controller
     public function store(Request $request)
 
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
 
         $validate = $request->validate([
             'date_start'=>'required',
@@ -169,7 +177,8 @@ class AgendaController extends Controller
      */
     public function edit($id)
     {
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
 
         $agenda = Agenda::find($id);
         $stands= Stand::all();
@@ -216,7 +225,9 @@ class AgendaController extends Controller
 
     public function destroy($id)
     {
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $agenda = Agenda::find($id);   
         if ($agenda) {
             $agenda->delete();
