+++ app/Http/Controllers/PlacesController.php
@@ -23,9 +23,13 @@ class PlacesController extends Controller
     private function userInauthenticated()
     {
         $this->user = $this->service->getUserAuthenticated();
-        if (!$this->user || $this->user->rol->nombre != 'Administrador') {
-            return view('auth/login', ['message' => 'No se ha logueado o no tiene los permisos']);
-        }
+        if (!$this->user) {
+            return view('auth/login', ['message' => 'No se ha logueado']);
+        } else if ($this->user->rol->name != 'Administrador') {
+            return view('home', ['message' => 'No tiene los permisos para ejecutar esta acción']);
+        } 
+        
+        return null;
     }
     /**
      * Display a listing of the resource.
@@ -34,7 +38,9 @@ class PlacesController extends Controller
      */
     public function index()
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $places= Places::with('schedule')->get();
         //dd($places);
         return view('places.index', compact('places'));
@@ -49,7 +55,9 @@ class PlacesController extends Controller
 
     public function create()    
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $schedules= Schedule::all();
         return view('places.create', compact('schedules'));
     }
@@ -62,7 +70,9 @@ class PlacesController extends Controller
      */
     public function store(Request $request)
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $validate = $request->validate([
             'name'=>'required',
             'email'=>'required',
@@ -102,7 +112,9 @@ class PlacesController extends Controller
      */
     public function edit($id)
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $place = Places::find($id);
         $schedules= Schedule::all();
         return view('places.edit', compact('place','schedules'));
@@ -117,7 +129,9 @@ class PlacesController extends Controller
      */
     public function update(Request $request, $id)
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $validate = $request->validate([
             'name'=>'required',
             'email'=>'required',
@@ -147,7 +161,9 @@ class PlacesController extends Controller
      */
     public function destroy($id)
     {   
-        $this->userInauthenticated();
+        $userInauthenticated = $this->userInauthenticated();
+        if ($userInauthenticated !== null) return $userInauthenticated;
+
         $place = Places::find($id);   
         if ($place) {
             $place->delete();
